Remove building key on all methods that were using it
Since adapter is used by stash-it, and stash-it builds key using adapter's buildKey method, having it built here produces doubled namespace as a result.
Because adapter is not meant to be used as a standalone tool and key is being built using stash-it, it should not be build here as well.